backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
newspaper3k==0.2.8
yfinance==0.2.54
pandas==2.2.3
numpy==2.1.3
pydantic==2.10.3
pytest==8.3.4
